+++ src/index.ts
@@ -9,9 +9,10 @@ import {
     getFrontend,
     getBackend,
     IModel,
-    Setting,
-    fetchPost,
-    Protyle, openWindow, IOperation
+    Protyle,
+    openWindow,
+    IOperation,
+    Constants
 } from "siyuan";
 import "@/index.scss";
 
@@ -120,6 +121,9 @@ export default class PluginSample extends Plugin {
                 text: "This is my custom dock"
             },
             type: DOCK_TYPE,
+            resize() {
+                console.log(DOCK_TYPE + " resize");
+            },
             init() {
                 this.element.innerHTML = `<div class="fn__flex-1 fn__flex-column">
     <div class="block__icons">
@@ -308,7 +312,7 @@ export default class PluginSample extends Plugin {
 
     private showDialog() {
         let dialog = new Dialog({
-            title: "Hello World",
+            title: `SiYuan ${Constants.SIYUAN_VERSION}`,
             content: `<div id="helloPanel" class="b3-dialog__content"></div>`,
             width: this.isMobile ? "92vw" : "720px",
             destroyCallback(options) {
@@ -527,6 +531,18 @@ export default class PluginSample extends Plugin {
                 label: "Off loaded-protyle-dynamic",
                 click: () => {
                     this.eventBus.off("loaded-protyle-dynamic", this.eventBusLog);
+                }
+            }, {
+                icon: "iconSelect",
+                label: "On switch-protyle",
+                click: () => {
+                    this.eventBus.on("switch-protyle", this.eventBusLog);
+                }
+            }, {
+                icon: "iconClose",
+                label: "Off switch-protyle",
+                click: () => {
+                    this.eventBus.off("switch-protyle", this.eventBusLog);
                 }
             }, {
                 icon: "iconSelect",